requirements.txt
python-dotenv
google-genai

google-cloud-speech
google-cloud-texttospeech
google-auth
google-auth-oauthlib
google-auth-httplib2

pymysql
sqlalchemy
fastapi
uvicorn

google-cloud-texttospeech pygame

passlib[bcrypt]
python-jose[cryptography]
email-validator